+++ src/tastytrade/utils/logging.py
@@ -1,6 +1,4 @@
 import logging
-from types import SimpleNamespace
-from typing import Any
 
 
 def setup_logging(level: int = logging.INFO) -> None:
@@ -11,8 +9,3 @@ def setup_logging(level: int = logging.INFO) -> None:
         datefmt="%Y-%m-%d %H:%M:%S",
         force=True,
     )
-
-
-def dict_to_class(data: dict[str, Any]) -> SimpleNamespace:
-    clean_data = {k.replace("-", "_"): v for k, v in data.items()}
-    return SimpleNamespace(**clean_data)
